Rename board to "/boards"
